+++ cellseg/data.py
@@ -1,11 +1,10 @@
 from torch.utils.data import Dataset
-from PIL import Image
+from skimage.io import imread
 import torch
 import glob
-import matplotlib.pyplot as plt
 
 
-class CellSeg(Dataset):
+class DataLoader(Dataset):
     def __init__(self, image_directory, image_suffix="tif"):
         self.image_directory = image_directory
         self.image_suffix = image_suffix
@@ -17,8 +16,12 @@ class CellSeg(Dataset):
     def __getitem__(self, img_index):
         if torch.is_tensor(img_index):
             img_index = img_index.tolist()
+        if self.image_suffix == "tif":
+            final_image = imread(self.image_list[img_index], plugin="pil")
+        else:
+            final_image = imread(self.image_list[img_index])
 
-        return {"image": Image.open(self.image_list[img_index])}
+        return {"image": final_image, "index": img_index}