+++ docs/source/changelog.rst
@@ -1,8 +1,16 @@
 
-cellseg: MultiClass Cell Segmentation
-=====================================
+cellseg: Quantify and Classify Cells
+====================================
 
 **version 0.1.0**
 
 
-* Conceptualize project 
+* 
+  Added ``show_images`` in ``utils.py`` to allow quick visualization of a given number of images from a given stack of
+  images. 
+
+* 
+  Implemented data loaders. 
+
+* 
+  Conceptualized project